+++ ft_strlcpy.c
@@ -1,9 +1,9 @@
 #include "libft.h"
 
-int	ft_strlcpy(char *dst, const char *src, int dstsize)
+size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	len;
-	int	i;
+	size_t	len;
+	size_t	i;
 
 	if (!src)
 		return (0);
